requirements.txt
pyjwt
Django==4.1.3
psycopg2-binary
djangorestframework
tzdata
python-dotenv